# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MatchTCG Backend API
  description: |
    MatchTCG is a backend API for Trading Card Game event management. It enables users to create, discover, and manage tournaments and casual matches with features like geospatial search, group management, calendar integration, and GDPR compliance.
    
    ## Authentication
    Most endpoints require authentication using Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```
    
    ## Rate Limiting
    - Authenticated users: 1000 requests/hour
    - Anonymous users: 100 requests/hour
    - Geospatial search: 60 requests/hour
    
    ## Error Handling
    All errors follow a consistent format with HTTP status codes and structured error responses.
  version: 1.0.0
  contact:
    name: MatchTCG API Support
    url: https://matchtcg.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.matchtcg.com/api/v1
    description: Production server
  - url: https://staging-api.matchtcg.com/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development server
security:
  - BearerAuth: []
  - {}
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: User successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using a refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token successfully refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate the current access token
      requestBody:
        required: false
      responses:
        '200':
          description: User successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully logged out
        '400':
          description: Invalid token format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/oauth/google:
    get:
      tags:
        - Authentication
      summary: Google OAuth authentication
      description: Initiate Google OAuth flow or handle callback
      security: []
      parameters:
        - name: code
          in: query
          description: OAuth authorization code (callback only)
          schema:
            type: string
        - name: state
          in: query
          description: OAuth state parameter (callback only)
          schema:
            type: string
      responses:
        '200':
          description: OAuth URL generated or callback processed
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      auth_url:
                        type: string
                        format: uri
                      state:
                        type: string
                  - $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid OAuth parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/oauth/apple:
    get:
      tags:
        - Authentication
      summary: Apple OAuth authentication
      description: Initiate Apple OAuth flow or handle callback
      security: []
      parameters:
        - name: code
          in: query
          description: OAuth authorization code (callback only)
          schema:
            type: string
        - name: state
          in: query
          description: OAuth state parameter (callback only)
          schema:
            type: string
      responses:
        '200':
          description: OAuth URL generated or callback processed
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      auth_url:
                        type: string
                        format: uri
                      state:
                        type: string
                  - $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid OAuth parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # User Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from authentication endpoints
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: validation_error
        message:
          type: string
          description: Human-readable error message
          example: Invalid input data
...
    AuthUserInfo:
      type: object
      required:
        - id
        - email
        - locale
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
          nullable: true
          example: "John Doe"
        locale:
          type: string
          enum: [en, pt]
          example: pt
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - timezone
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "securepassword123"
        display_name:
          type: string
          maxLength: 100
          example: "John Doe"
        locale:
          type: string
          enum: [en, pt]
          example: pt
        timezone:
          type: string
          example: "Europe/Lisbon"
        country:
          type: string
          example: "Portugal"
        city:
          type: string
          example: "Lisbon"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "securepassword123"
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
    AuthResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - expires_at
        - user
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_at:
          type: string
          format: date-time
          example: "2024-01-01T15:30:00Z"
        user:
          $ref: '#/components/schemas/AuthUserInfo'

    # User Management Schemas
...